+++ packages/router-plugin/src/router.state.ts
@@ -1,4 +1,4 @@
-import { NgZone, Injectable, OnDestroy } from '@angular/core';
+import { NgZone, Injectable, OnDestroy, Injector } from '@angular/core';
 import {
   NavigationCancel,
   NavigationError,
@@ -6,15 +6,12 @@ import {
   RouterStateSnapshot,
   RoutesRecognized,
   ResolveEnd,
-  UrlSerializer,
   NavigationStart,
-  NavigationEnd
+  NavigationEnd,
+  Event
 } from '@angular/router';
-import { LocationStrategy, Location } from '@angular/common';
 import { Action, Selector, State, StateContext, Store } from '@ngxs/store';
-import { isAngularInTestMode } from '@ngxs/store/internals';
 import { Subscription } from 'rxjs';
-import { first } from 'rxjs/operators';
 
 import {
   Navigate,
@@ -22,9 +19,16 @@ import {
   RouterCancel,
   RouterError,
   RouterNavigation,
-  RouterDataResolved
+  RouterDataResolved,
+  RouterRequest,
+  RouterNavigated
 } from './router.actions';
 import { RouterStateSerializer } from './serializer';
+import {
+  NavigationActionTiming,
+  NgxsRouterPluginOptions,
+  NGXS_ROUTER_PLUGIN_OPTIONS
+} from './symbols';
 
 export interface RouterStateModel<T = RouterStateSnapshot> {
   state?: T;
@@ -32,13 +36,12 @@ export interface RouterStateModel<T = RouterStateSnapshot> {
   trigger: RouterTrigger;
 }
 
-export type RouterTrigger = 'none' | 'router' | 'store';
-
-/**
- * @description Will be provided through Terser global definitions by Angular CLI
- * during the production build. This is how Angular does tree-shaking internally.
- */
-declare const ngDevMode: boolean;
+export type RouterTrigger =
+  | 'none'
+  | 'router'
+  | 'store'
+  // The `devtools` trigger means that the state change has been triggered by Redux DevTools (e.g. when the time-traveling is used).
+  | 'devtools';
 
 @State<RouterStateModel>({
   name: 'router',
@@ -66,9 +69,11 @@ export class RouterState implements OnDestroy {
    */
   private _storeState: RouterStateModel | null = null;
 
-  private _lastRoutesRecognized: RoutesRecognized = null!;
+  private _lastEvent: Event | null = null;
 
   private _subscription = new Subscription();
+
+  private _options: NgxsRouterPluginOptions | null = null;
 
   @Selector()
   static state<T = RouterStateSnapshot>(state: RouterStateModel<T>) {
@@ -85,13 +90,13 @@ export class RouterState implements OnDestroy {
     private _router: Router,
     private _serializer: RouterStateSerializer<RouterStateSnapshot>,
     private _ngZone: NgZone,
-    private _urlSerializer: UrlSerializer,
-    private _locationStrategy: LocationStrategy,
-    private _location: Location
+    injector: Injector
   ) {
-    this.setUpStoreListener();
-    this.setUpRouterEventsListener();
-    this.checkInitialNavigationOnce();
+    // Note: do not use `@Inject` since it fails on lower versions of Angular with Jest
+    // integration, it cannot resolve the token provider.
+    this._options = injector.get(NGXS_ROUTER_PLUGIN_OPTIONS, null);
+    this._setUpStoreListener();
+    this._setUpRouterEventsListener();
   }
 
   ngOnDestroy(): void {
@@ -108,99 +113,123 @@ export class RouterState implements OnDestroy {
     );
   }
 
-  @Action([RouterNavigation, RouterError, RouterCancel, RouterDataResolved])
+  @Action([
+    RouterRequest,
+    RouterNavigation,
+    RouterError,
+    RouterCancel,
+    RouterDataResolved,
+    RouterNavigated
+  ])
   angularRouterAction(
     ctx: StateContext<RouterStateModel>,
     action: RouterAction<RouterStateModel, RouterStateSnapshot>
   ): void {
     ctx.setState({
-      ...ctx.getState(),
       trigger: action.trigger,
       state: action.routerState,
       navigationId: action.event.id
     });
   }
 
-  private setUpStoreListener(): void {
+  private _setUpStoreListener(): void {
     const subscription = this._store
       .select(RouterState)
       .subscribe((state: RouterStateModel | undefined) => {
-        this.navigateIfNeeded(state);
+        this._navigateIfNeeded(state);
       });
 
     this._subscription.add(subscription);
   }
 
-  private setUpRouterEventsListener(): void {
+  private _navigateIfNeeded(routerState: RouterStateModel | undefined): void {
+    if (routerState && routerState.trigger === 'devtools') {
+      this._storeState = this._store.selectSnapshot(RouterState);
+    }
+
+    const canSkipNavigation =
+      !this._storeState ||
+      !this._storeState.state ||
+      !routerState ||
+      routerState.trigger === 'router' ||
+      this._router.url === this._storeState.state.url ||
+      this._lastEvent instanceof NavigationStart;
+
+    if (canSkipNavigation) {
+      return;
+    }
+
+    this._storeState = this._store.selectSnapshot(RouterState);
+    this._trigger = 'store';
+    this._ngZone.run(() => this._router.navigateByUrl(this._storeState!.state!.url));
+  }
+
+  private _setUpRouterEventsListener(): void {
+    const dispatchRouterNavigationLate =
+      this._options != null &&
+      this._options.navigationActionTiming === NavigationActionTiming.PostActivation;
+
+    let lastRoutesRecognized: RoutesRecognized;
+
     const subscription = this._router.events.subscribe(event => {
+      this._lastEvent = event;
+
       if (event instanceof NavigationStart) {
-        this.navigationStart();
+        this._navigationStart(event);
       } else if (event instanceof RoutesRecognized) {
-        this._lastRoutesRecognized = event;
+        lastRoutesRecognized = event;
+        if (!dispatchRouterNavigationLate && this._trigger !== 'store') {
+          this._dispatchRouterNavigation(lastRoutesRecognized);
+        }
       } else if (event instanceof ResolveEnd) {
-        this.dispatchRouterDataResolved(event);
+        this._dispatchRouterDataResolved(event);
       } else if (event instanceof NavigationCancel) {
-        this.dispatchRouterCancel(event);
-        this.reset();
+        this._dispatchRouterCancel(event);
+        this._reset();
       } else if (event instanceof NavigationError) {
-        this.dispatchRouterError(event);
-        this.reset();
+        this._dispatchRouterError(event);
+        this._reset();
       } else if (event instanceof NavigationEnd) {
-        this.navigationEnd();
-        this.reset();
+        if (this._trigger !== 'store') {
+          if (dispatchRouterNavigationLate) {
+            this._dispatchRouterNavigation(lastRoutesRecognized);
+          }
+          this._dispatchRouterNavigated(event);
+        }
+        this._reset();
       }
     });
 
     this._subscription.add(subscription);
   }
 
-  private navigationStart(): void {
+  /** Reacts to `NavigationStart`. */
+  private _navigationStart(event: NavigationStart): void {
     this._routerState = this._serializer.serialize(this._router.routerState.snapshot);
 
     if (this._trigger !== 'none') {
       this._storeState = this._store.selectSnapshot(RouterState);
+      this._dispatchRouterAction(new RouterRequest(this._routerState, event, this._trigger));
     }
   }
 
-  private navigationEnd(): void {
-    if (this.shouldDispatchRouterNavigation()) {
-      this.dispatchRouterNavigation();
-    }
+  /** Reacts to `ResolveEnd`. */
+  private _dispatchRouterDataResolved(event: ResolveEnd): void {
+    const routerState = this._serializer.serialize(event.state);
+    this._dispatchRouterAction(new RouterDataResolved(routerState, event, this._trigger));
   }
 
-  private shouldDispatchRouterNavigation(): boolean {
-    if (!this._storeState) return true;
-    return this._trigger !== 'store';
-  }
+  /** Reacts to `RoutesRecognized` or `NavigationEnd`, depends on the `navigationActionTiming`. */
+  private _dispatchRouterNavigation(lastRoutesRecognized: RoutesRecognized): void {
+    const nextRouterState = this._serializer.serialize(lastRoutesRecognized.state);
 
-  private navigateIfNeeded(state: RouterStateModel | undefined): void {
-    const canSkipNavigation =
-      !this._storeState ||
-      !this._storeState.state ||
-      !state ||
-      state.trigger === 'router' ||
-      this._router.url === this._storeState.state.url;
-
-    if (canSkipNavigation) {
-      return;
-    }
-
-    this._trigger = 'store';
-    this._ngZone.run(() => {
-      this._router.navigateByUrl(this._storeState!.state!.url);
-    });
-  }
-
-  private dispatchRouterNavigation(): void {
-    const nextRouterState = this._serializer.serialize(this._lastRoutesRecognized.state);
-
-    this.dispatchRouterAction(
+    this._dispatchRouterAction(
       new RouterNavigation(
         nextRouterState,
         new RoutesRecognized(
-          this._lastRoutesRecognized.id,
-          this._lastRoutesRecognized.url,
-          this._lastRoutesRecognized.urlAfterRedirects,
+          lastRoutesRecognized.id,
+          lastRoutesRecognized.url,
+          lastRoutesRecognized.urlAfterRedirects,
           nextRouterState
         ),
         this._trigger
@@ -208,14 +237,16 @@ export class RouterState implements OnDestroy {
     );
   }
 
-  private dispatchRouterCancel(event: NavigationCancel): void {
-    this.dispatchRouterAction(
+  /** Reacts to `NavigationCancel`. */
+  private _dispatchRouterCancel(event: NavigationCancel): void {
+    this._dispatchRouterAction(
       new RouterCancel(this._routerState!, this._storeState, event, this._trigger)
     );
   }
 
-  private dispatchRouterError(event: NavigationError): void {
-    this.dispatchRouterAction(
+  /** Reacts to `NavigationEnd`. */
+  private _dispatchRouterError(event: NavigationError): void {
+    this._dispatchRouterAction(
       new RouterError(
         this._routerState!,
         this._storeState,
@@ -225,7 +256,13 @@ export class RouterState implements OnDestroy {
     );
   }
 
-  private dispatchRouterAction<T>(action: RouterAction<T>): void {
+  /** Reacts to `NavigationEnd`. */
+  private _dispatchRouterNavigated(event: NavigationEnd): void {
+    const routerState = this._serializer.serialize(this._router.routerState.snapshot);
+    this._dispatchRouterAction(new RouterNavigated(routerState, event, this._trigger));
+  }
+
+  private _dispatchRouterAction<T>(action: RouterAction<T>): void {
     this._trigger = 'router';
 
     try {
@@ -235,67 +272,9 @@ export class RouterState implements OnDestroy {
     }
   }
 
-  private dispatchRouterDataResolved(event: ResolveEnd): void {
-    const routerState = this._serializer.serialize(event.state);
-    this.dispatchRouterAction(new RouterDataResolved(routerState, event, this._trigger));
-  }
-
-  private reset(): void {
+  private _reset(): void {
     this._trigger = 'none';
     this._storeState = null;
     this._routerState = null;
-  }
-
-  /**
-   * No sense to mess up the `setUpRouterEventsListener` method as we have
-   * to perform this check only once and unsubscribe after the first event
-   * is triggered
-   */
-  private checkInitialNavigationOnce(): void {
-    // Caretaker note: we have still left the `typeof` condition in order to avoid
-    // creating a breaking change for projects that still use the View Engine.
-    if (
-      (typeof ngDevMode === 'undefined' || ngDevMode) &&
-      // Angular is running tests in development mode thus we can be sure that this method will be
-      // skipped in tests.
-      isAngularInTestMode()
-    ) {
-      return;
-    }
-
-    const subscription = this._router.events
-      .pipe(first((event): event is RoutesRecognized => event instanceof RoutesRecognized))
-      .subscribe(({ url }) => {
-        // `location.pathname` always equals manually entered URL in the address bar
-        // e.g. `location.pathname === '/foo'`, but the `router` state has been initialized
-        // with another URL (e.g. used in combination with `NgxsStoragePlugin`), thus the
-        // `RouterNavigation` action will be dispatched and the user will be redirected to the
-        // previously saved URL. We want to prevent such behavior, so we perform this check
-
-        // `url` is a recognized URL by the Angular's router, while `currentUrl` is an actual URL
-        // entered in the browser's address bar
-        // `PathLocationStrategy.prototype.path()` returns a concatenation of
-        // `PlatformLocation.pathname` and normalized `PlatformLocation.search`.
-
-        // `Location.prototype.normalize` strips base href from the URL,
-        // if `baseHref` (declared in angular.json) for example is `/en`
-        // and the URL is `/test#anchor` - then `_locationStrategy.path(true)` will return `/en/test#anchor`,
-        // but `/en/test#anchor` is not known to the Angular's router, so we have to strip `/en`
-        // from the URL
-        const currentUrl = this._location.normalize(this._locationStrategy.path(true));
-        const currentUrlTree = this._urlSerializer.parse(currentUrl);
-        // We need to serialize the URL because in that example `/test/?redirect=https://google.com/`
-        // Angular will recognize it as `/test?redirect=https:%2F%2Fwww.google.com%2F`
-        // so we have to run the `currentUrl` via the `UrlSerializer` that will encode characters
-        const currentSerializedUrl = this._urlSerializer.serialize(currentUrlTree);
-
-        // If URLs differ from each other - we've got to perform a redirect to the manually entered URL
-        // in the address bar, as it must have a priority
-        if (currentSerializedUrl !== url) {
-          this._router.navigateByUrl(currentUrl);
-        }
-      });
-
-    this._subscription.add(subscription);
   }
 }