+++ packages/store/operators/src/compose.ts
@@ -1,8 +1,10 @@
-import { StateOperator } from '@ngxs/store';
-import { NoInfer } from './utils';
+import { ExistingState, NoInfer, StateOperator } from './types';
 
 export function compose<T>(...operators: NoInfer<StateOperator<T>[]>): StateOperator<T> {
-  return function composeOperator(existing: Readonly<T>): T {
-    return operators.reduce((accumulator, operator) => operator(accumulator), existing);
+  return function composeOperator(existing: ExistingState<T>): T {
+    return operators.reduce(
+      (accumulator, operator) => operator(accumulator as ExistingState<T>),
+      existing as T
+    );
   };
 }