+++ packages/store/src/internal/lifecycle-state-manager.ts
@@ -1,75 +1,98 @@
-import { Injectable } from '@angular/core';
-import { NgxsBootstrapper, PlainObject } from '@ngxs/store/internals';
-import { filter, mergeMap, tap } from 'rxjs/operators';
+import { Injectable, OnDestroy } from '@angular/core';
+import { NgxsBootstrapper } from '@ngxs/store/internals';
+import { EMPTY, Subject } from 'rxjs';
+import {
+  catchError,
+  filter,
+  mergeMap,
+  pairwise,
+  startWith,
+  takeUntil,
+  tap
+} from 'rxjs/operators';
 
+import { Store } from '../store';
+import { getValue } from '../utils/utils';
+import { InternalErrorReporter } from './error-handler';
 import { StateContextFactory } from './state-context-factory';
 import { InternalStateOperations } from './state-operations';
-import { getStateDiffChanges, MappedStore, StatesAndDefaults } from './internals';
+import { MappedStore, StatesAndDefaults } from './internals';
 import { NgxsLifeCycle, NgxsSimpleChange, StateContext } from '../symbols';
 
 @Injectable()
-export class LifecycleStateManager {
+export class LifecycleStateManager implements OnDestroy {
+  private readonly _destroy$ = new Subject<void>();
+
   constructor(
-    private internalStateOperations: InternalStateOperations,
-    private stateContextFactory: StateContextFactory,
-    private bootstrapper: NgxsBootstrapper
+    private _store: Store,
+    private _internalErrorReporter: InternalErrorReporter,
+    private _internalStateOperations: InternalStateOperations,
+    private _stateContextFactory: StateContextFactory,
+    private _bootstrapper: NgxsBootstrapper
   ) {}
 
+  ngOnDestroy(): void {
+    this._destroy$.next();
+  }
+
   ngxsBootstrap<T>(action: T, results: StatesAndDefaults | undefined): void {
-    this.internalStateOperations
+    this._internalStateOperations
       .getRootStateOperations()
       .dispatch(action)
       .pipe(
         filter(() => !!results),
-        tap(() => this.invokeInit(results!.states)),
-        mergeMap(() => this.bootstrapper.appBootstrapped$),
-        filter(appBootstrapped => !!appBootstrapped)
+        tap(() => this._invokeInitOnStates(results!.states)),
+        mergeMap(() => this._bootstrapper.appBootstrapped$),
+        filter(appBootstrapped => !!appBootstrapped),
+        catchError(error => {
+          // The `SafeSubscriber` (which is used by most RxJS operators) re-throws
+          // errors asynchronously (`setTimeout(() => { throw error })`). This might
+          // break existing user's code or unit tests. We catch the error manually to
+          // be backward compatible with the old behavior.
+          this._internalErrorReporter.reportErrorSafely(error);
+          return EMPTY;
+        }),
+        takeUntil(this._destroy$)
       )
-      .subscribe(() => this.invokeBootstrap(results!.states));
+      .subscribe(() => this._invokeBootstrapOnStates(results!.states));
   }
 
-  /**
-   * Invoke the init function on the states.
-   */
-  invokeInit(mappedStores: MappedStore[]): void {
+  private _invokeInitOnStates(mappedStores: MappedStore[]): void {
     for (const mappedStore of mappedStores) {
       const instance: NgxsLifeCycle = mappedStore.instance;
 
       if (instance.ngxsOnChanges) {
-        const currentAppState: PlainObject = {};
-        const newAppState: PlainObject = this.internalStateOperations
-          .getRootStateOperations()
-          .getState();
-
-        const firstDiffChange: NgxsSimpleChange = getStateDiffChanges(mappedStore, {
-          currentAppState,
-          newAppState
-        });
-
-        instance.ngxsOnChanges(firstDiffChange);
+        this._store
+          .select(state => getValue(state, mappedStore.path))
+          .pipe(startWith(undefined), pairwise(), takeUntil(this._destroy$))
+          .subscribe(([previousValue, currentValue]) => {
+            const change = new NgxsSimpleChange(
+              previousValue,
+              currentValue,
+              !mappedStore.isInitialised
+            );
+            instance.ngxsOnChanges!(change);
+          });
       }
 
       if (instance.ngxsOnInit) {
-        instance.ngxsOnInit(this.getStateContext(mappedStore));
+        instance.ngxsOnInit(this._getStateContext(mappedStore));
       }
 
       mappedStore.isInitialised = true;
     }
   }
 
-  /**
-   * Invoke the bootstrap function on the states.
-   */
-  invokeBootstrap(mappedStores: MappedStore[]) {
+  private _invokeBootstrapOnStates(mappedStores: MappedStore[]) {
     for (const mappedStore of mappedStores) {
       const instance: NgxsLifeCycle = mappedStore.instance;
       if (instance.ngxsAfterBootstrap) {
-        instance.ngxsAfterBootstrap(this.getStateContext(mappedStore));
+        instance.ngxsAfterBootstrap(this._getStateContext(mappedStore));
       }
     }
   }
 
-  private getStateContext(mappedStore: MappedStore): StateContext<any> {
-    return this.stateContextFactory.createStateContext(mappedStore);
+  private _getStateContext(mappedStore: MappedStore): StateContext<any> {
+    return this._stateContextFactory.createStateContext(mappedStore);
   }
 }