+++ packages/store/src/ivy/ivy-enabled-in-dev-mode.ts
@@ -1,5 +1,3 @@
-import { ɵivyEnabled } from '@angular/core';
-
 import { getUndecoratedStateInIvyWarningMessage } from '../configs/messages.config';
 
 /**
@@ -12,11 +10,9 @@ export function ensureStateClassIsInjectable(stateClass: any): void {
   // AOT mode because this property is added before runtime. If an application is running in
   // JIT mode then this property can be added by the `@Injectable()` decorator. The `@Injectable()`
   // decorator has to go after the `@State()` decorator, thus we prevent users from unwanted DI errors.
-  if (ɵivyEnabled) {
-    const ngInjectableDef = stateClass.ɵprov;
-    if (!ngInjectableDef) {
-      // Don't warn if Ivy is disabled or `ɵprov` exists on the class
-      console.warn(getUndecoratedStateInIvyWarningMessage(stateClass.name));
-    }
+  const ngInjectableDef = stateClass.ɵprov;
+  if (!ngInjectableDef) {
+    // Don't warn if Ivy is disabled or `ɵprov` exists on the class
+    console.warn(getUndecoratedStateInIvyWarningMessage(stateClass.name));
   }
 }