+++ packages/store/src/private_api.ts
@@ -1,0 +1,2 @@
+export { NgxsRootModule as ɵNgxsRootModule } from './modules/ngxs-root.module';
+export { NgxsFeatureModule as ɵNgxsFeatureModule } from './modules/ngxs-feature.module';