+++ packages/store/src/store.ts
@@ -1,13 +1,13 @@
 // tslint:disable:unified-signatures
 import { Inject, Injectable, Optional, Type } from '@angular/core';
-import { Observable, of, Subscription, throwError } from 'rxjs';
+import { Observable, of, Subscription, throwError, queueScheduler } from 'rxjs';
 import {
   catchError,
   distinctUntilChanged,
   map,
-  publishReplay,
-  refCount,
-  take
+  shareReplay,
+  take,
+  observeOn
 } from 'rxjs/operators';
 import { INITIAL_STATE_TOKEN, PlainObject } from '@ngxs/store/internals';
 
@@ -28,9 +28,9 @@ export class Store {
    * All selects would use this stream, and it would call leave only once for any state change across all active selectors.
    */
   private _selectableStateStream = this._stateStream.pipe(
+    observeOn(queueScheduler),
     leaveNgxs(this._internalExecutionStrategy),
-    publishReplay(1),
-    refCount()
+    shareReplay({ bufferSize: 1, refCount: true })
   );
 
   constructor(