+++ packages/store/src/symbols.ts
@@ -2,9 +2,11 @@ import { Injectable, InjectionToken, Type } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { PlainObject, StateClass } from '@ngxs/store/internals';
-import { SharedSelectorOptions } from './internal/internals';
-import { NgxsExecutionStrategy } from './execution/symbols';
+import { StateOperator } from '@ngxs/store/operators';
+
 import { DispatchOutsideZoneNgxsExecutionStrategy } from './execution/dispatch-outside-zone-ngxs-execution-strategy';
+import { NgxsExecutionStrategy } from './execution/symbols';
+import { SharedSelectorOptions } from './internal/internals';
 import { StateToken } from './state-token/state-token';
 
 export const ROOT_STATE_TOKEN = new InjectionToken<any>('ROOT_STATE_TOKEN');
@@ -31,8 +33,9 @@ export class NgxsConfig {
    * - Object.freeze on the state and actions to guarantee immutability
    * (default: false)
    *
-   * @deprecated This property is no longer necessary when the Ivy compiler is used.
-   * We'll determine the development mode through the `ngDevMode`. It's still essential with View Engine.
+   * Note: this property will be accounted only in development mode when using the Ivy compiler.
+   * It makes sense to use it only during development to ensure there're no state mutations.
+   * When building for production, the Object.freeze will be tree-shaken away.
    */
   developmentMode: boolean;
   compatibility: {
@@ -78,7 +81,7 @@ export class NgxsConfig {
   }
 }
 
-export type StateOperator<T> = (existing: Readonly<T>) => T;
+export { StateOperator };
 
 /**
  * State context provided to the actions in the state.
@@ -153,7 +156,7 @@ export class NgxsSimpleChange<T = any> {
  * On init interface
  */
 export interface NgxsOnInit {
-  ngxsOnInit(ctx?: StateContext<any>): void | any;
+  ngxsOnInit(ctx: StateContext<any>): void;
 }
 
 /**
@@ -167,7 +170,7 @@ export interface NgxsOnChanges {
  * After bootstrap interface
  */
 export interface NgxsAfterBootstrap {
-  ngxsAfterBootstrap(ctx?: StateContext<any>): void;
+  ngxsAfterBootstrap(ctx: StateContext<any>): void;
 }
 
 export type NgxsModuleOptions = Partial<NgxsConfig>;